Modify tree-sitter descend-to-child behavior
If descending does not encounter any nodes with siblings on the way to
the leaf, then, instead of visiting the leaf, change it to not descend
at all, to be consistent with "select nearest" behavior which selects
the outermost rather than the innermost node. This avoids getting
stuck in preorder traversals in some cases due to point-height offset
calculation restoring the cursor location to the outermost node where
we can still "descend" to a child node even though this outer node
would not be visited using tree navigation commands directly.

In addition, we also now visit nodes without siblings if they start at
a different position in the buffer than the parent. E.g. in the
statement `import datetime`, with the above change to not visit
leaves, we would not visit `datetime` without this additional change.