Register repeatable commands
Evil uses a command property `:repeat` to determine whether to record
repeat information for any given function. This adds a list of commands
which should have their `:repeat` function set to `t` which signals to
evil that the keystrokes associated with each should be recorded to
`evil-repeat-ring.` This should apply to all symex commands which edit
the buffer but do not leave the symex state.