Do not repeat commands w/o :repeat prop
The `evil-repeat-type` command takes a DEFAULT argument which is used
when a function doesn't have any value for :repeat. Within evil, this
function is given `t` as a default value because evil mode repeats all
commands by default and only ignores ones that are specifically marked
as repeatable.

In symex, the opposite is true. Repeatable commands are explicitly
defined and any function without a :repeat value should be ignored.