Fix bug when quoting and unquoting prefixes overlap
E.g. if # is used as a quote prefix, then its use in an unquote prefix
like #,@ or #, would not be correctly removed using
symex-delete-prefix, which would only remove # instead. This fixes it
so that quoting and unquoting prefixes are merged before deciding
which one to delete (it deletes the longest from this combined list,
as before).