fix #49: Make move-y direction customizable (and reverse default)
This commit implements the feature as requested in #49. We imagine the root to
be level 0, so that going UP a level corresponds with going DOWN the tree. To
eliminate the confusion the function names are changed to go-in/out.

We do not change the functionality of the functions (that would be confusing),
but merely flip the keybindings (except for the hydra, no time to set bindings
conditionally).

As suggested in one of the comments this commit reverses the
default direction.

Aliases are created for reverse compatibility.