TS: add symex-ts--handle-tree-modification, use during paste
The new `symex-ts--handle-tree-modification` macro is an attempt to
generalise how tree modifications are handled with respect (mainly) to
selecting a new current node.

This macro first saves the current tree, performs the transformation
and then compares the newly modified tree. It attempts to find the
first range that is marked as being modified and, if successful, moves
the point to the start of that range. Regardless of the outcome, it
then uses the point to select the current node and handles
mode-specific indentation.

`symex-ts--paste` has been modified to use this new macro. The hope is
that this macro will also prove useful when implementing other
transformations.