ts: select neighboring text before attempting to select node
In some cases setting the node (e.g. on entry into Symex) would result
in selecting the entire file tree if point is positioned within
whitespace (e.g. on a blank line). This changes it so that we first
position point on a nearby non-whitespace character before attempting
to set the current node on entry.