Eliminate side effects as a special parameter in the evaluator
Formerly, side effects could be specified as a mode of
evaluation. Instead, we now just have them as a distinct traversal
type. This allows us to attach side effects to any traversal or even
components of traversals without needing to treat them in a special
way. This simplifies the evaluator and gives us more flexibility in
general.