Fixes in Symex's computation model
This simplifies Symex's fold-like, "monadic" comptutation model. A
`nil` result from `symex-execute-traversal` (eventually to be renamed
`symex-eval`) implies failure, while anything else (including a zero
move) is an arbitrary computation done on the executed
traversal. Although formerly there were a lot of idiosyncratic but
perhaps cool things planned with how the computation would work, this
now just fine tunes the basic model which is:

- lifting the individual moves in the traversal into the computation
  "type"
- composing these lifted computations to get the computation result

The default computation, as before, is to return a list representing
the sequence of executed moves.