(ts) reselect node from point if it's outdated
Register a notifier to be notified of buffer changes, and each time,
check if the current node is "outdated." If it is, then reselect
current node from point.

This should ensure that Symex always points to a valid current node
after any changes to the buffer, and we shouldn't need to wrap
individual transformation (i.e., mutating) commands anymore.