set source major mode in transform-in-isolation macro
Formerly we were setting the syntax table from the original buffer,
but for some reason, this doesn't work in some cases, and motions
in the macro body don't consistently behave the same as they do in the
source buffer (e.g., comments are treated as regular code in some
cases). But even more importantly, it wouldn't work if we used any
treesitter-specific code, as the behavior of those is dependent on the
major mode and attendant parsers, and not just the syntax table.