fix point-height offset for initial node in a buffer
Now that we don't allow the root to be selected, the initial node is
the "topmost" selectable node and its point-height offset is 0 rather
than 1.

This was causing the initial symex to be unselectable, and was
thus affecting most operations done in a temporary buffer using the
"transform in isolation" macro, since the expressions to be
transformed are always at the initial position in such buffers.