update node selection after transformation
This avoids encountering an "outdated node" error when we attempt a
motion after the transformation, by resetting the selection to a valid
node post-transformation.

Another option could be to ensure, in the primitive motions, that the
current node is not outdated, and update the selection from point if
it is.