do offset calculations in source buffer for shift-forward
Unlike Lisp, where the basic syntax is explicit and not
context-sensitive, in treesitter (e.g., Python), the same code
extracted from the source buffer and pasted into a temporary buffer
(which we are doing in some transformations, to do them "in
isolation," as that in principle makes things simpler) might produce a
different parse tree. For instance, what is a "dotted_name" in the
source buffer could, without the containing context being available,
be considered an "expression_statement" in the temporary buffer.

This was causing computed offsets (e.g., start and end of node) to be
different in the source vs the temporary buffer, leading to the result
of the transformation sometimes being unexpected.

By doing the offset calculations in the source buffer first, and then
doing the transformation in the temporary buffer using these values,
we avoid such issues.