handle whitespace and separator differently for computing end point
Formerly we lumped these two into a single flag `include-whitespace`,
which controlled whether the end point of the symex included trailing
whitespace and also any trailing separator like a comma. But with
"open line below," we need to include the trailing comma but *not* any
trailing whitespace, so we now distinguish these with separate flags.

This also refactors to eliminate an unnecessary intermediary
function (`symex-remove`).

Note: for some reason, in a Clojure buffer, this inserts an extra
newline. It might have something to do with `forward-char` apparently
being overridden to `rect-forward-char` in such buffers, but that's
just a wild guess.