refactor symex-misc: introduce new "symex-tree" module
This module contains tree-level insights and abilities, that is,
typically things that use the Symex evaluator to compute things about
the tree, either in providing new abilities like leaping, or insights
like position on a certain branch, remaining length, height above the
root. It also contains the implementation of "branch memory," which
remembers positions on branches when going up and down.