use "ordinary" major modes when transforming in isolation
The `symex--transform-in-isolation` macro transforms text using a
temporary buffer in the same major mode as the source buffer, to
ensure that syntax (e.g., comments) and operations (e.g., forward by
one symex) have the same meaning as in the source buffer. But if the
source buffer happens to be "special" in some way, such as being a
REPL, then it could lead to unexpected behavior due to such major
modes exhibiting special behavior.

We now identify such special cases (currently just IELM, the ELisp
REPL) and use the corresponding "ordinary" major mode (e.g.,
`emacs-lisp-mode`) to perform the transformation.