refactor `paste` for uniformity across treesitter and lisp
Refactor to use the treesitter implementation of paste in abstracted
form for both lisp and treesitter. Formerly, we were already doing
this uniformly using the DSL, and I initially started changing it
because I mistakenly thought that there were two separate
implementations of yank and paste in the code. But the DSL
implementation hadn't been thoroughly tested in treesitter, and since
this new implementation is working OK in these cases, we can just
retain it for now unless we discover a good reason to change it back.