Move branch memory inside motions
Formerly this was being done by advising motions (i.e., "advice").
Treating it as part of the motion, instead, simplifies things a bit
and keeps the concerns (like post-selection side effects) in one
place. This is also a step towards treating returning to a branch
position formally as part of the traversal (e.g., reflected in the
return value representing the executed traversal).