properly handle precaution failure
When a traversal guarded by a precaution fails, it was still executing
the "afterwards" condition, which, if that succeeded, then it was
considering the traversal to be a success. Yet, what we actually want
is for the traversal to fail if either of the conditions (pre/post)
fails *or* if the traversal itself fails. This changes it to only
check the post-condition if the traversal succeeds.