Upgrade branch memory
Formerly, this feature would work well in the (most common) case of
moving up and down the tree just one step at a time. But if we ever
used a count argument, the memory would no longer be accurate. This
was just ephemeral and easily shaken off as any lateral motion resets
the memory, so it didn't cause any lasting confusion.

We now leverage branch memory within the traversals to go up and down,
instead of extrinsically as an add-on feature, so that it now works
with arbitrary counts.