don't select nearest symex in "transform in isolation" macro
We don't need to, as, in Lisp, it will already be at the right spot,
and in treesitter, the buffer change notifier would ensure that the
nearest symex is selected. Not depending on this primitive binding
allows us to move this utility into the lowest "utils" layer, so that
it can be used even in Lisp/TS primitives.