restore DSL implementation of paste
The other, more recent, implementation had some unhandled edge cases
when used with counts. The former DSL implementation was accidentally
replaced as I didn't realize it was already implemented in a common
way for both Lisp and treesitter. It is a reasonably mature and
well-tested implementation and seems to work well. The former
redundancy with unused implementations of paste which had caused the
confusion is no longer present now after recent refactorings.