restore height offset in shift backward/forward
In treesitter, if we shift at the start of a block, it was selecting
the entire block after the operation since nearest selection just
selects the "highest" node at point with any context awareness. But as
we are in the midst of shifting expressions at a certain level, we
would like to remain at that level after the operation. This change
ensures that we stay at the same hierarchy level after the operation
by restoring the "point-height offset," the same way we do in the
`symex-save-excursion` macro.