don't always select nearest if a symex is already selected
In some cases, we do need to reset the selection based on point, but
in other cases, we just need to ensure that there is a valid
selection. This changes a couple of places where we really need to do
the latter, but where it was doing the former. This avoids reseting a
valid state to an invalid selection in the context of an operation.
For instance, in shifting an expression at the start of a block in
treesitter, it was formerly resetting the selection after the
operation as well as tidying, as a matter of course. This was causing
the context of the operation, and the desire to stay at the same level
as the transformed expression, to be lost, as nearest selection was
selecting the "topmost" node at point, which was the entire block
rather than just the first expression.