use `symex-save-excursion` for uniformity
We need to return to the original height offset after transformations,
as that information could be lost during the transformation. For
example, tidying a line could cause the node selection on that line to
become stale, which should trigger resetting of the selection based on
point, which selects the "topmost" node at point. But as, in
treesitter, nodes at different hierarchy levels may have the same
starting point, the "height offset" of the initially selected node is
lost. We were manually storing and restoring that in the tidy and
shift-forward operations, but it looks like we can just use the
`symex-save-excursion` macro which basically does the same thing.