revert use of `symex-save-excursion`
This change was recently made as it seemed equivalent to what was
there before, but looks like it isn't quite the same and introduces
weird behavior in Lisp where it hangs on attempting to indent after a
paste, and also doesn't select the right symex after pasting, in
certain cases. Correct functioning is now hopefully restored.