remove `clojure-ts-mode` from `symex-lisp-modes`
It looks like this major mode changes the syntax table so that the
usual Emacs motions for symexes no longer work the same way. So, it
seems, using the treesitter parser is more appropriate for this major
mode, after all. Really it seems to come down to choosing between
`clojure-mode` and `clojure-ts-mode` for use with Clojure. With Symex,
the former uses the Lisp parser, and the latter, the treesitter
parser. Currently, the former provides a more consistent experience.