make effects lambdas instead of expressions
This would allow us to consider eventually making these actually
_traversals_, which as a base case can be lambdas. But currently, we
don't go that far because this would require effect functions to
return a truthy value in order to avoid the traversal failure logic
that restores point to its original location and so on, which we may
not want to do for such effects and might lead to some weird edge
cases. Generally, when using the effect form, it's most likely that we
will be using a simple lambda rather than a complicated traversal, so
we might as well just support this case for now.