improve handling of counts in going up/down with memory
Formerly, attempting to go up or down N times where N exceeds the
actual distance to the original position or to the root caused the
stack to be fully popped and the positions to be lost. We now handle
this case and restore the original stack in the event the traversal
fails.

Note: this points the way forward on implementing predicates within
the DSL as traversals instead of entering ELisp for this, as that
would simplify these traversals.