Allow DSL conditions to be traversals
This generalizes the conditions used in traversals like precaution,
decision, and loop, to be traversals themselves (while still
supporting the existing syntactic sugar of `at`, `and`, `root`, etc.).
This allows these forms to condition on the result of traversal
execution. The base case for any traversal is now reinstated as being
any ELisp function, including a lambda, with the requirement that this
lambda accept the same arguments as any traversal, i.e., the
computation to be performed and the in-progress accumulated result.

In practice, traversals that are lambdas are likely to just be
conditions, and, as such, their result will not be integrated into the
computation. They will thus typically ignore the computation argument,
and as long as they return a truthy or a nil result, they can serve
their purpose (as they do today in precautions and so on) without
needing to know how to compose computations, which traversals
typically must do, and, often, without even using the accumulated
traversal result.