Repeat all Symex commands by default, exclude non-transformations
Formerly, we were recording only symex "commands," i.e., which are
transformations, and we were not recording motions. This stopped
working once we moved to Lithium because when used through this
interface, the value of `this-command` no longer reflects as the Symex
commands we've registered for repetition, but instead, as the wrapping
lambdas used in Lithium. We do modify `this-command` in these wrapping
lambdas in Lithium to make this seamless in general, but that doesn't
help in this case since repetition is tracked in *pre-command* hooks,
where we haven't yet changed the value of `this-command`.

The workaround now is to _always_ record commands while in Symex mode,
and specifically exclude the ones we don't want to be repeatable,
e.g., motions. This isn't an ideal solution, and there are reasons to
prefer the more explicit version we had before (mentioned in
comments).

In cases where a command entails exiting Symex mode (e.g.,
`wrap-and-insert` or simply `i`) when using `evil-repeat-pop`, it
appears to leave things in a weird state when used with Rigpa, where
it enters Symex evil state (used for UI feedback only) even though it
isn't actually in Symex (lithium) mode. So, some further testing, or
perhaps, a simpler and more robust solution (like a generic (not
evil-specific) repeat package), is needed.