remove "blink cursor" customization
This can be done in user config via hooks, and it may be preferable to
just disable the cursor outright (also in user config).

We'll avoid maintaining customizations that are too specialized and
which aren't directly related to Symex specifically, in favor of just
recommending recipes for such desired customization in user config via
the provided lifecycle hooks.