handle the use of a completion menu in repetition in a general way
This now simply ignores any activity that has no effect in the source
buffer, whether that's to insert or delete text or even simply move
point. This is the case while navigating completion menus, where we
use C-n and C-p but that neither changes the source buffer nor moves
point.

Ignoring activity that has no effect seems like a good practice for
repetition, and it may be more general than completion.