Ensure lithium mode is enabled
In addition to just using Lithium to define the modal interface, it's
also necessary for lithium-mode to be enabled, as it ensures that the
necessary hooks are in place that allow Lithium modes to play well
with the minibuffer, among other things.

Formerly, it was assumed this would be part of user config, but it's
better for Symex to ensure this, too, as it would in fact be
preferable to avoid having distinct config for lithium if the user is
only using it for Symex.

Fixes #156