ts: set the current node on selection
We weren't officially setting the current node in some cases upon
selecting the nearest node. It's possible this would have had
unexpected behavior in some cases, such as performing a transformation
immediately upon entry to Symex mode in this particular selection
configuration, but I didn't test it. Just noticed this discrepancy and
changed it to reflect intent.