remove lisp/treesitter overrides as they probably don't do anything
IIUC the usual major modes for languages like clojure do *not* use
treesitter, and only the dedicated treesitter modes do. So, overriding
the usual ones to use the Lisp parser isn't necessary. In cases where
the major mode uses treesitter, I think it also means that the syntax
table is modified and uses a treesitter interpretation, so that
overriding with the Lisp parser wouldn't have the same behavior as in
a non-treesitter major mode.