add symex-evil and symex-rigpa modules
The former was originally part of symex but was moved to a separate
repo as part of making it a separate package, but that's not necessary
to do. Instead it will be better to keep these modules in the common
symex repo and just register them as different packages within it,
while taking care to keep the dependencies isolated in some
form (ideally keeping them in separate folders).