ci: bump low version for testing up to 29.1
Technically, as we are guarding all calls to treesit* APIs, it should
still work on older versions of Emacs, just, without treesitter
support (e.g., you could still use it with Lisp languages).

But attempts to tell the linter and bytecompiler to ignore these
functions on older versions of Emacs (via, e.g., `declare-function`)
didn't seem to have any effect, so we'll just bump the CI version
tested so that the results are accurately reported (as far as we
know).

The actual dependency in the Symex `Package-Requires` header remains
`25.1`, so we still support installing Symex on older versions of
Emacs.