parse insertions uniformly, including those that initiate parsing
`I` and `A` (for instance) implicitly insert spaces before entering
the insertion mode. As these changes are reported to the buffer change
listener before the command completes, the parser doesn't already
include this entered key sequence (in its mapped, parsed form), and we
therefore parse it as an ordinary key sequence rather than as an
insertion. This is fine in practice, but it introduces some incidental
complexity as we would prefer to handle all insertions uniformly ---
i.e., parsing them as insertions rather than as key sequences.

This simply changes the condition for capturing buffer changes to
check either that parsing is already in progress, or it is being
initiated by the current key.