parse insertion along two axes: inserted text and point motion
The inserted text also now includes an offset, so that it could record
insertions at any position relative to point.

This is towards fixing a bug with implicit insertion of text such as
matching delimiters (e.g., with paredit), where the repetition inserts
the text in the order of insertion without regard to placement in
relation to point. This now handles both insertion and placement
correctly.