(repeat) parse initial key as key sequence
The initial key may sometimes entail buffer changes, e.g., `I` inserts
whitespace prior to entering the insertion state.

While we usually parse buffer insertions as insertions, in this case,
we prefer to parse it as a key sequence, since otherwise, the
remaining parsed activity would presumably be in the context of
`symex-editing-mode` which it would never have left, rather than an
insertion state. There could be many ways of handling this, but
parsing the initial key as a key sequence seems simplest.

At an earlier stage, this was being done by reflecting on the parser
state in the `map` predicate, but that was causing a native
compilation warning as this predicate was one of the components of the
parser definition, so there was some circularity there. We now no
longer refer to the parser itself in any of these component
predicates.