remove offset and point motion from parsing of insertions
This was causing a regression where repeating actions involving
completion menus was no longer working (i.e., this fixes #171).

There is at least one bug in the removed logic, which is that
deletions should also translate `symex--replaying-point` (in addition
to insertions).

This logic was introduced specifically to handle capturing insertions
from ( and [ that don't insert directly at point. But now we handle
those simply as key sequences, so this logic is present purely for
hypothetical cases.

Dealing with offsets and point motion increases the complexity of
repeat parsing. As there isn't a known application for it at the
present time, for now, we will just remove it and restore proper
functionality on known cases, and thereby also simplify parsing. We
can always reintroduce if we encounter a tangible unhandled case that
cannot be handled by an alternative strategy such as parsing it as a
key sequence or as a command.