Fix capture of insertions for repetition via `.` (dot)
For key sequences that (1) don't move point, and (2) don't insert any
text in the buffer, we were capturing them as *no-ops*.
But *initiating keys* like `i` which change modes (i.e., enters an
insertion state like vanilla Emacs or Evil's insert state), while
being no-ops from that perspective, still need to be parsed as key
sequences since they change the meaning of subsequent keys.

To fix it, we simply change the priority of parsing the intiating key
to be higher than checking if it's a no-op.

Fixes #174.