fix repeating paste during insertion with `.` (dot)
In some cases (e.g., evidently when using paste (`C-y` in vanilla
Emacs) while inserting), some of the reported changes from that action
are being considered neither insertions nor deletions, which were
being parsed as `nil`. This is not a recognized mantra, so Symex mode
was defensively exiting when the resulting error was encountered.

We now parse these as `mantra--null`, the proper null value to use
when we mean that the mantra is a no-op (the actual insertion
associated with the paste is being captured separately).

Related to #174.